# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgenticNews API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ArticleTrustScore =
  | (typeof ArticleTrustScore)[keyof typeof ArticleTrustScore]
  | null;

export const ArticleTrustScore = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface TimelineEvent {
  date: string;
  event: string;
}

export type TrustDetailsScore =
  (typeof TrustDetailsScore)[keyof typeof TrustDetailsScore];

export const TrustDetailsScore = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface TrustDetails {
  score: TrustDetailsScore;
  reasoning: string;
  sources: string[];
  contradictions: string[];
}

export interface Article {
  id: number;
  title: string;
  description: string;
  content: string;
  url: string;
  imageUrl?: string | null;
  source: string;
  author?: string | null;
  publishedAt: string;
  category: string;
  aiSummary?: string | null;
  bulletPoints?: string[] | null;
  timeline?: TimelineEvent[] | null;
  trustScore?: ArticleTrustScore;
  trustDetails?: TrustDetails | null;
  audioUrl?: string | null;
  createdAt: string;
}

export interface NewsResponse {
  articles: Article[];
  total: number;
  page: number;
  pageSize: number;
}

export interface SummaryResponse {
  summary: string;
  bulletPoints: string[];
  timeline: TimelineEvent[];
}

export type VerificationResponseTrustScore =
  (typeof VerificationResponseTrustScore)[keyof typeof VerificationResponseTrustScore];

export const VerificationResponseTrustScore = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface VerificationResponse {
  trustScore: VerificationResponseTrustScore;
  reasoning: string;
  sources: string[];
  contradictions: string[];
}

export interface AudioResponse {
  audioUrl: string;
  message: string;
}

export interface TranslateRequest {
  targetLanguage: string;
}

export interface TranslateResponse {
  translatedTitle: string;
  translatedContent: string;
  language: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface AskRequest {
  question: string;
  conversationHistory?: ChatMessage[];
}

export interface ArticleReference {
  id: number;
  title: string;
  source: string;
  url: string;
}

export interface AskResponse {
  answer: string;
  sources: ArticleReference[];
}

export interface UploadResponse {
  extractedText: string;
  summary: string;
  bulletPoints: string[];
  insights: string;
}

export interface CategoriesResponse {
  categories: string[];
}

export type GetNewsParams = {
  /**
   * News category
   */
  category?: string;
  /**
   * Search query
   */
  q?: string;
  page?: number;
  pageSize?: number;
};

export type UploadArticleBody = {
  image: Blob;
};