# lib/api-zod/src/generated/types/askRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgenticNews API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";

export interface AskRequest {
  question: string;
  conversationHistory?: ChatMessage[];
}