# lib/api-zod/src/generated/types/askResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgenticNews API
 * OpenAPI spec version: 0.1.0
 */
import type { ArticleReference } from "./articleReference";

export interface AskResponse {
  answer: string;
  sources: ArticleReference[];
}