# lib/api-zod/src/generated/types/newsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgenticNews API
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from "./article";

export interface NewsResponse {
  articles: Article[];
  total: number;
  page: number;
  pageSize: number;
}