# lib/api-zod/src/generated/types/summaryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgenticNews API
 * OpenAPI spec version: 0.1.0
 */
import type { TimelineEvent } from "./timelineEvent";

export interface SummaryResponse {
  summary: string;
  bulletPoints: string[];
  timeline: TimelineEvent[];
}