# lib/api-zod/src/generated/types/trustDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgenticNews API
 * OpenAPI spec version: 0.1.0
 */
import type { TrustDetailsScore } from "./trustDetailsScore";

export interface TrustDetails {
  score: TrustDetailsScore;
  reasoning: string;
  sources: string[];
  contradictions: string[];
}