# lib/api-zod/src/generated/types/uploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgenticNews API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResponse {
  extractedText: string;
  summary: string;
  bulletPoints: string[];
  insights: string;
}