# lib/api-zod/src/generated/types/verificationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgenticNews API
 * OpenAPI spec version: 0.1.0
 */
import type { VerificationResponseTrustScore } from "./verificationResponseTrustScore";

export interface VerificationResponse {
  trustScore: VerificationResponseTrustScore;
  reasoning: string;
  sources: string[];
  contradictions: string[];
}